+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.8.3
 
 * Fix rails hooking for version 7.1. [#327](https://github.com/bogdan/datagrid/issues/327)
+* Fix formatting of value for `date` and `datetime-local` input types
 
 ## 1.8.2